+++ Samples/Metrics.Samples/SampleMetrics.cs
@@ -21,7 +21,7 @@ namespace Metrics.Samples
         /// <summary>
         /// keep a histogram of the input data of our requet method 
         /// </summary>
-        private readonly Histogram histogramOfData = Metric.Histogram("ResultsExample", Unit.Items, SamplingType.LongTerm);
+        private readonly Histogram histogramOfData = Metric.Histogram("ResultsExample", Unit.Items);
 
         /// <summary>
         /// measure the rate at which requests come in
@@ -31,7 +31,7 @@ namespace Metrics.Samples
         /// <summary>
         /// measure the time rate and duration of requests
         /// </summary>
-        private readonly Timer timer = Metric.Timer("Requests", Unit.Requests, SamplingType.FavourRecent, TimeUnit.Seconds, TimeUnit.Milliseconds);
+        private readonly Timer timer = Metric.Timer("Requests", Unit.Requests);
 
         private double someValue = 1;
 
@@ -46,7 +46,7 @@ namespace Metrics.Samples
             new MultiContextMetrics().Run();
             MultiContextInstanceMetrics.RunSample();
 
-            using (this.timer.NewContext()) // measure until disposed
+            using (this.timer.NewContext(i.ToString())) // measure until disposed
             {
                 someValue *= (i + 1); // will be reflected in the gauge 
 
@@ -56,7 +56,7 @@ namespace Metrics.Samples
 
                 this.meter.Mark(); // signal a new request to the meter
 
-                this.histogramOfData.Update(ThreadLocalRandom.NextLong() % 5000); // update the histogram with the input data
+                this.histogramOfData.Update(ThreadLocalRandom.NextLong() % 5000, i.ToString()); // update the histogram with the input data
 
 
                 // simulate doing some work