+++ Src/Metrics/Core/DefaultMetricsBuilder.cs
@@ -1,6 +1,7 @@
 ﻿
 using System;
 using Metrics.PerfCounters;
+using Metrics.Sampling;
 namespace Metrics.Core
 {
     public sealed class DefaultMetricsBuilder : MetricsBuilder