+++ Src/Metrics/Gauge.cs
@@ -6,8 +6,8 @@ namespace Metrics
     /// </summary>
     public sealed class GaugeValueSource : MetricValueSource<double>
     {
-        public GaugeValueSource(string name, MetricValueProvider<double> value, Unit unit)
-            : base(name, value, unit)
+        public GaugeValueSource(string name, MetricValueProvider<double> value, Unit unit, MetricTags tags)
+            : base(name, value, unit, tags)
         { }
     }
 }