+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+###0.2.2 / 2014-09-27
+* add support for tagging metrics (not yet used in reports or visualization)
+* add support for suppling a string user value to histograms & timers for tracking min / max / last values
+* tests cleanup, some refactoring
+
 ###0.2.1 / 2014-09-25
 * port latest changes from original metrics lib
 * port optimization from ExponentiallyDecayingReservoir (https://github.com/etishor/Metrics.NET/commit/1caa9d01c16ff63504612d64771d52e9d7d9de5e)